# fix deeplink highlighting