before:
```
 go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/keybase/client/go/chat/utils
BenchmarkDecorateLinks-4              24          49435327 ns/op
PASS
ok      github.com/keybase/client/go/chat/utils 1.667s
```

after:
```
$ go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/keybase/client/go/chat/utils
BenchmarkDecorateLinks-4              21          54543743 ns/op
PASS
ok      github.com/keybase/client/go/chat/utils 1.618s
```